requirements.txt
streamlit==1.32.0         # GUI
pandas==2.0.3            # Data handling
scikit-learn==1.3.0      # ML model
gensim==4.3.2            # Word embeddings
nltk==3.8.1              # NLP processing
python-dotenv==1.0.0     # API keys
Pillow==9.5.0  # Recommended stable version
numpy==1.26.0  # exact version